mdcode: keep BigQuery DDL assertions out of the loader test (#4)
loader.test.ts had an end-to-end block that imported the BigQuery generator
and asserted DDL fragments (KEY/MEASURE/REFERENCES) — destination-specific
tests in the loader's unit file. Move that block (and its live-verified
two-hop fan-out fixture) to bigquery.e2e.test.ts, where the BigQuery
destination is tested. loader.test.ts now imports only the loader and asserts
only the IR; raw YAML/JSON text parsing stays covered by its document-level
tests. No test lost: 74 pass across 3 files.
